fix: Move common styles in PortalModal to component/common/styles.js
